feat(Mapster): Feito a configuração e implementação do Mapster para mapeamento das Entidades/Dtos da aplicação.

+++ FinanceEdgeTrack/Domain/Models/Receita.cs
@@ -1,4 +1,5 @@
 ﻿using FinanceEdgeTrack.Domain.Models.Abstract;
+using Mapster;
 using System.ComponentModel.DataAnnotations;
 
 namespace FinanceEdgeTrack.Domain.Models;
@@ -7,6 +8,6 @@ public class Receita : Categoria
 {
     [Required(ErrorMessage = "É obrigatório informar o valor da receita.")]
     [Range(typeof(decimal), "1", "999999999999")]
-    public double Valor { get; set; }
+    public decimal Valor { get; set; }
     public DateTime Data { get; set; }
 }
